backend/requirements.txt
anyascii==0.3.1
audioread==3.0.0
Babel==2.12.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coqpit==0.0.17
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.7
decorator==5.1.1
docopt==0.6.2
Flask==2.2.3
fonttools==4.39.0
fsspec==2023.3.0
g2pkk==0.1.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==1.2.0
kiwisolver==1.4.4
librosa==0.8.0
llvmlite==0.38.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mecab-python3==1.0.5
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
platformdirs==3.1.1
playsound==1.3.0
pooch==1.7.0
protobuf==3.19.6
psutil==5.9.4
pycparser==2.21
pydub==0.25.1
pynndescent==0.5.8
pyparsing==3.0.9
pypinyin==0.48.0
pysbd==0.3.4
pytesseract==0.3.10
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
tensorboardX==2.6
threadpoolctl==3.1.0
torch==1.13.1
torchaudio==0.13.1
tqdm==4.65.0
trainer==0.0.20
TTS==0.11.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.15
Wand==0.6.11
Werkzeug==2.2.3
zipp==3.15.0
pyoxidizer==0.24.0
